+++ tests/testthat/test-crew_user.R
@@ -1,0 +1,5 @@
+crew_test("crew_user()", {
+  skip_on_cran()
+  user <- crew_user()
+  expect_true(is.null(user) || is.character(user))
+})